requirements.txt
ultralytics
torch
torchvision
opencv-python
Pillow
matplotlib
streamlit
numpy
pandas
